+++ CHANGELOG.md
@@ -6,3 +6,4 @@ master
 
 * Add monolog processor to encrypt logs
 * Add optional sonata admin to decrypt logs
+* Add processor to decrypt secrets at runtime